Add files via upload
Here is the solidworks file for the sensor holder.  The extruded circles have been made slightly smaller than the diameters of the sensors to ensure a flex fit.  The cuts in the circles are not a measured amount.  Please double-check that the width for hooking onto the fish tank is indeed large enough.  Everything is measured in inches.